KARL: restore script dir searching
For.... reasons?  the MounRiver people "fixed" the search paths?  I
don't know what they were trying to do, but this breaks installations.
If you use "make install" and the bin ends up on a path that you expect,
then the scripts are in expected locations, and you don't need any "-s"
arguments.  Put it back how it was before MRS got their hands on it.

Fixes: a82b704157 Import MounRiver WCH fork
Change-Id: If48aef04659f591e0b04788d11131761126440ef
Signed-off-by: Karl Palsson <karlp@tweak.au>