target/cortex_m: check core implementor field
Presently, we only look at the Part Number field of the CPUID, and
completely ignore the Implmentor field, simply assuming it to be ARM.

Parts have since been found, with different implementors, that use
overlapping part numbers, causing detection to fail.

Include checks for the core implementor, and open up the part number to
be a plain integer, as we can no longer enumerate them.

Signed-off-by: Karl Palsson <karlp@tweak.au>